@@ -14,26 +14,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.9",
-    "@angular/common": "^6.1.9",
-    "@angular/compiler": "^6.1.9",
-    "@angular/core": "^6.1.9",
-    "@angular/forms": "^6.1.9",
-    "@angular/http": "^6.1.9",
-    "@angular/platform-browser": "^6.1.9",
-    "@angular/platform-browser-dynamic": "^6.1.9",
-    "@angular/platform-server": "^6.1.9",
-    "@angular/router": "^6.1.9",
-    "angular2-recaptcha": "^0.4.1",
+    "@angular/animations": "^7.2.16",
+    "@angular/common": "^7.2.16",
+    "@angular/compiler": "^7.2.16",
+    "@angular/core": "^7.2.16",
+    "@angular/forms": "^7.2.16",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^7.2.16",
+    "@angular/platform-browser-dynamic": "^7.2.16",
+    "@angular/platform-server": "^7.2.16",
+    "@angular/router": "^7.2.16",
+    "angular2-recaptcha": "^1.1.0",
     "core-js": "^2.5.1",
     "rxjs": "^6.3.3",
+    "tslib": "^1.9.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.5",
-    "@angular/cli": "^6.2.3",
-    "@angular/compiler-cli": "^6.1.9",
-    "@angular/language-service": "^6.1.9",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.10",
+    "@angular/compiler-cli": "^7.2.16",
+    "@angular/language-service": "^7.2.16",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
@@ -51,6 +52,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "^2.9.2"
+    "typescript": "~3.2.4"
   }
 }
