@@ -1,4 +1,4 @@
-import { Component } from "@angular/core";
+import { AfterViewInit, Component } from "@angular/core";
 import { Router, Event, NavigationEnd } from '@angular/router';
 import { GLOBALS } from "./shared/globals";
 
@@ -11,21 +11,29 @@ declare let ga:Function;
     styleUrls: ['./app.component.css']
 })
 
-export class AppComponent {
-    // Variables from external TS.
-    arrGlobals = GLOBALS;
-    
-    constructor(public router:Router) {
-        this.router.events.subscribe(
-            (event:Event) => {
-                if (event instanceof NavigationEnd) {
-                    // ga('send', 'pageview', event.urlAfterRedirects);
+export class AppComponent implements AfterViewInit {
+  // Variables from external TS.
+  arrGlobals = GLOBALS;
 
-                    // Updated as of 1/17/2017
-                    ga('set', 'page', event.urlAfterRedirects);
-                    ga('send', 'pageview');
-                }
-            }
-        ); // this.router.events.subscribe
-    } // constructor(public router:Router)
+  constructor(public router: Router) {
+
+  } // constructor(public router:Router)
+
+  ngAfterViewInit(): void {
+    this.router.events.subscribe(
+      (event: Event) => {
+        if (event instanceof NavigationEnd) {
+          // ga('send', 'pageview', event.urlAfterRedirects);
+
+          // Updated as of 1/17/2017
+          try {
+            ga('set', 'page', event.urlAfterRedirects);
+            ga('send', 'pageview');
+          } catch (error) {
+            console.log('GA loaded late.');
+          }
+        }
+      }
+    ); // this.router.events.subscribe
+  } // ngAfterViewInit()
 }
